requirements.txt
cassiopeia >= 0.1.3
discord.py >= 0.16.0